Renamed module `avl_tree` to `create_avl_tree`
