+++ ergotree-interpreter/src/eval/create_avl_tree.rs
@@ -3,9 +3,9 @@ use crate::eval::env::Env;
 use crate::eval::EvalContext;
 use crate::eval::EvalError;
 use ergotree_ir::chain::digest32::ADDigest;
-use ergotree_ir::mir::avl_tree::CreateAvlTree;
 use ergotree_ir::mir::avl_tree_data::{AvlTreeData, AvlTreeFlags};
 use ergotree_ir::mir::constant::TryExtractInto;
+use ergotree_ir::mir::create_avl_tree::CreateAvlTree;
 use ergotree_ir::mir::value::Value;
 use ergotree_ir::util::AsVecU8;
 use std::convert::TryFrom;
