+++ ergotree-ir/src/serialization/expr.rs
@@ -6,7 +6,6 @@ use crate::has_opcode::HasStaticOpCode;
 use crate::mir::and::And;
 use crate::mir::apply::Apply;
 use crate::mir::atleast::Atleast;
-use crate::mir::avl_tree::CreateAvlTree;
 use crate::mir::bin_op::ArithOp;
 use crate::mir::bin_op::BitOp;
 use crate::mir::bin_op::LogicalOp;
@@ -32,6 +31,7 @@ use crate::mir::collection::coll_sigma_parse;
 use crate::mir::collection::coll_sigma_serialize;
 use crate::mir::constant::Constant;
 use crate::mir::constant::ConstantPlaceholder;
+use crate::mir::create_avl_tree::CreateAvlTree;
 use crate::mir::create_prove_dh_tuple::CreateProveDhTuple;
 use crate::mir::create_provedlog::CreateProveDlog;
 use crate::mir::decode_point::DecodePoint;
